Implement domain renewal failure handling and add payment intent action webhook
