+++ apps/web/app/(ee)/api/stripe/webhook/route.ts
@@ -9,6 +9,7 @@ import { checkoutSessionCompleted } from "./checkout-session-completed";
 import { customerSubscriptionDeleted } from "./customer-subscription-deleted";
 import { customerSubscriptionUpdated } from "./customer-subscription-updated";
 import { invoicePaymentFailed } from "./invoice-payment-failed";
+import { paymentIntentRequiresAction } from "./payment-intent-requires-action";
 
 const relevantEvents = new Set([
   "charge.succeeded",
@@ -18,6 +19,7 @@ const relevantEvents = new Set([
   "customer.subscription.updated",
   "customer.subscription.deleted",
   "invoice.payment_failed",
+  "payment_intent.requires_action",
 ]);
 
 // POST /api/stripe/webhook – listen to Stripe webhooks
@@ -65,6 +67,9 @@ export const POST = async (req: Request) => {
         break;
       case "invoice.payment_failed":
         await invoicePaymentFailed(event);
+        break;
+      case "payment_intent.requires_action":
+        await paymentIntentRequiresAction(event);
         break;
     }
   } catch (error) {
